Update README.md
As i have completed my second and third task of Codsoft Internship, this is my Forth task which is a small game of Rock-Paper-Scissor with the computer. As we enter our choice from given one it will show choice of computer and make a decision that is it correct or not.